+++ lib/dns/zone/rr/https.rb
@@ -1,0 +1,7 @@
+# `HTTPS` resource record.
+#
+# RFC 9460. Shares the presentation format of {DNS::Zone::RR::SVCB}, so parsing
+# and dumping are inherited.
+class DNS::Zone::RR::HTTPS < DNS::Zone::RR::SVCB
+
+end